bucket: split bucket tests into 3 files (normal, list and manager)
